pyproject.toml
[project]
name = "dayforge"
version = "0.1.0"
description = "DayForge — a personal life planner for schedules, habits, and learning goals"
requires-python = ">=3.11"
dependencies = [
    "psycopg[binary,pool]>=3.2",
    "python-dotenv>=1.0",
]

[project.scripts]
dayforge = "server:main"

[dependency-groups]
dev = [
    "ruff>=0.9",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
